Initial graph generation for the data reporting.
Extends the report page to also generate a graph on the same page.  This
will be cleaned up for presentation in the future.
Added the ProjectData interface for reporting.  This is a generic
interface that will allow various data element strategies to be created
to generate various graphs.
Initial implementation of the OpenVsResolved strategy class that will
examine the issues within a project and provide a breakdown of status
over the given time span.